Refactor Controller
Landing is now render by React, so comment .render so as not to interfere. This need to be done correctly, this is just a patch.

And, data from the ML API it's not sent directly to the front (pug), it´s stored in a .json file, than the front will pick up.